viewmodel 수정 완료
