--- lib/data/repository/sign_up_repository/sign_up_repository.dart
@@ -1,5 +1,0 @@
-abstract class SignUpRepository {
-  bool nicknameVaildate(String value);
-  bool isFullyValidKoreanOrEnglish(String value);
-  String getHelperText(String value);
-}
